refactor(ls): remove unnecessary targetSpec fields
Currently we only support OpenAPI 3.1.0 and 3.0.x versions.
Every rule (documentation, completion, linting) that is not
explicitly annotated with targetSpec is assumed to be the same
for both specification versions.

If in future, we introduce OpenAPI 2.0, we would first need
to annotate all the common rules explicitly with union
targetSpecs of OpenAPI 3.1.0 and 3.0.x.