fix(core): fix serialization of BooleanElement(false) elements
When BooleanElement(false) get's converted to primitive
false value, this value was returned from visitor method.
Visitor mechanism uses boolean false value returned from
visitor methods as indicator that it should skip visiting
the node.
Thus the value that is used as an indicator is now configurable
and for the purpose of toValue serialization, different
indicator is used.

Refs https://github.com/swagger-api/swagger-editor/issues/3723