fix(core): fix isArrayElement predicate polymorhic characteristics
As ObjectElement is extension of ArrayElement,
isArrayElement returned true for every ObjectElement instance.
Although that was technically correct, it was casing
nasty bugs when using isArrayElement.

This commit make the fact that the ObjectElement is extension
of ArrayElement an implementation detail, and from now on
isArrayElement returns false when asserting on ObjectElement.