fix(apidom-ls): Add array node handling in mustache/handlebars utils
Extended the handlebar utilities function with an array node checking condition which will return the first value of an array node if it exists.
The basic test templates and related files have been also updated for testing these changes.
A getContext function was added to the context-short.ts file as an exported function for greater functionality.